refactor: dedupe helpers — merge subscription_checker into task_manager; single parse_limit; shared is_subscription_active/content_totals/work_loads_summary; reuse pyro formatters in /stats
