+++ Backend/fastapi/routes/api_routes.py
@@ -40,8 +40,7 @@ from Backend.pyrofork.bot import (
 async def get_system_stats_api():
     try:
         db_stats = await db.get_database_stats()
-        total_movies = sum(stat.get("movie_count", 0) for stat in db_stats)
-        total_tv_shows = sum(stat.get("tv_count", 0) for stat in db_stats)
+        total_movies, total_tv_shows = db.content_totals(db_stats)
         api_tokens = await db.get_all_api_tokens()
         
         return {
@@ -228,25 +227,28 @@ async def delete_tv_season_api(tmdb_id: int, db_index: int, season: int):
 
 
 #----- Token management
+#----- Parse a GB-limit value into a positive float, or None
+def _parse_limit(val):
+    try:
+        v = float(val)
+        return v if v > 0 else None
+    except (ValueError, TypeError, AttributeError):
+        return None
+
+
 async def create_token_api(payload: dict):
     try:
         token_name = payload.get("name")
         daily_limit = payload.get("daily_limit_gb")
         monthly_limit = payload.get("monthly_limit_gb")
-        
+
         if not token_name:
-             raise HTTPException(status_code=400, detail="Token name is required")
-        def parse_limit(val):
-            try:
-                v = float(val)
-                return v if v > 0 else None
-            except (ValueError, TypeError):
-                return None
+            raise HTTPException(status_code=400, detail="Token name is required")
 
         new_token = await db.add_api_token(
-            token_name, 
-            parse_limit(daily_limit), 
-            parse_limit(monthly_limit)
+            token_name,
+            _parse_limit(daily_limit),
+            _parse_limit(monthly_limit)
         )
         return new_token
     except Exception as e:
@@ -256,18 +258,11 @@ async def update_token_limits_api(token: str, payload: dict):
     try:
         daily_limit = payload.get("daily_limit_gb")
         monthly_limit = payload.get("monthly_limit_gb")
-        
-        def parse_limit(val):
-            try:
-                v = float(val)
-                return v if v > 0 else None
-            except (ValueError, TypeError, AttributeError):
-                return None
 
         await db.update_api_token_limits(
             token,
-            parse_limit(daily_limit),
-            parse_limit(monthly_limit)
+            _parse_limit(daily_limit),
+            _parse_limit(monthly_limit)
         )
         return {"message": "Limits updated successfully"}
